Update tests
Update test to adjust to the new API. Refactor test creating test
suites.
