+++ src/test/java/com/burgerjavis/mvc/BurgerJavisServerMvcTest.java
@@ -1,0 +1,15 @@
+/**
+ * Diego Moreno Lennon
+ * Burger Javi's Server
+ */
+
+package com.burgerjavis.mvc;
+
+import org.junit.runner.RunWith;
+import org.junit.runners.Suite;
+
+@RunWith(Suite.class)
+@Suite.SuiteClasses({})
+public class BurgerJavisServerMvcTest {
+
+}
